+++ src/main/kotlin/io/github/dylmeadows/eontimer/util/Calibrations.kt
@@ -4,14 +4,14 @@ import io.github.dylmeadows.eontimer.model.settings.Console
 
 object Calibrations {
 
-    fun convertToMillis(delays: Int, console: Console): Long =
-        Math.round(delays * console.frameRate)
+    fun convertToMillis(delays: Int, console: Console): Int =
+        Math.round(delays * console.frameRate).toInt()
 
-    fun convertToDelays(millis: Int, console: Console): Long =
-        Math.round(millis / console.frameRate)
+    fun convertToDelays(millis: Int, console: Console): Int =
+        Math.round(millis / console.frameRate).toInt()
 
-    fun createCalibration(delay: Int, second: Int, console: Console): Long {
-        val delayCalibration = delay - convertToDelays(second * 1000, console).toInt()
+    fun createCalibration(delay: Int, second: Int, console: Console): Int {
+        val delayCalibration = delay - convertToDelays(second * 1000, console)
         return convertToMillis(delayCalibration, console)
     }
 }
